sistemato l'errore di esecuzione dopo la scrittura della prima pagina. Ora è da sistemare la struttura dati dell'addrspace in modo da non dover aprire ogni volta il file dell'header in quanto QUESTO CAUSA L'INVALIDAZIONE DELLA TLB. In pratica non salvo mai niente nella tlb. Deve essere salvato nell'addrspace il vnode, così da non doverlo aprire ogni volta. In as destroy invece va fatta la distruzione del vnode invece che del progname. In runprogram va sostituito il codice dove si copia il nome del programma ma invece va salvato il vnode dentro l'addrspace.

+++ os161-base-2.0.3/kern/include/segments.h
@@ -22,7 +22,7 @@ struct segment{
 
 struct addrspace;
 
-int load_page(struct addrspace* as, int npage, paddr_t paddr, uint8_t segment);
+int load_page(struct addrspace* as, int npage, paddr_t paddr, int segment);
 
 
 #endif //_SEGMENTS_H_
